requirements.txt
altgraph==0.17
attrs==19.3.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
constantly==15.1.0
crypto==1.4.1
cryptography==3.0
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
demjson==2.2.4
entrypoints==0.3
et-xmlfile==1.0.1
Flask==1.1.1
future==0.18.2
html5lib==1.1
hyperlink==20.0.1
idna==2.8
importlib-metadata==1.7.0
incremental==17.5.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itemadapter==0.1.0
itemloaders==1.0.2
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
jieba==0.42.1
Jinja2==2.10.3
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
Naked==0.1.31
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.19.0
openpyxl==3.0.4
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parsel==1.6.0
parso==0.7.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
Protego==0.1.16
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.2
PyInstaller==3.6
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
queuelib==1.5.0
requests==2.24.0
retry==0.9.2
Scrapy==2.3.0
selenium==2.48.0
Send2Trash==1.5.0
service-identity==18.1.0
shellescape==3.8.1
six==1.15.0
soupsieve==2.0.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
Twisted==20.3.0
urllib3==1.25.7
virtualenv==16.7.9
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wordcloud==1.8.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
zope.interface==5.1.0
